Move initSubmenu JS files around, fix link issue and IE11 issue
